Add an example conversion test.
The webhook hookup in suite_test.go does hook up and configure the mutating
and validating webhooks.  It also registers the conversion webhook, and this
can be seen in the log messages.  However, it does not install the config
for the conversion webhook--this requires that kustomize add a conversion
webhook config to the CRDs.  There are comments in
envtest.WebhookInstallOptions that indicate that it can be used for this, but
I still have to dig into how to use it.