Conversion library and test from clusterapi
Now we can actually do *loss-less* conversions between the v1alpha1 and
v1alpha2 versions of the Desert and Vehicle resources.

The conversion library contains routines to marshal/unmarshal the up-rev
version of the resource via an annotation it places on the down-rev version.
The library contains its own test as well.

Use the conversion library in api/v1alpha1/conversion.go.

Add test api/v1alpha1/conversion_test.go for conversions between the hub,
v1alpha2, and the spoke, v1alpha1.  The test uses the test functions that come
with the conversion library, and uses a fuzzing library to do an exhaustive
test of the conversion routines without running through the conversion webhook.

Add simpler conversion tests in internal/controller/*_controller_test.go.
These tests confirm that the conversion webhook uses the conversion routines.

Enable ginkgo arg processing in api/v1alpha1/conversion.go and in the
conversion library's own test.  This is because we like to add ginkgo args in
the "make test" target, and ginkgo has to be loaded to recognize those args,
even if the test module isn't otherwise using ginkgo.